+++ tests/testthat/test-basic.r
@@ -123,3 +123,32 @@ test_that('nested module can use parent', {
     expect_true(exists('z', b))
     expect_equal(b$z, 1)
 })
+
+test_that('using legacy functions raises warning', {
+    on.exit({
+        box::unload(library)
+        box::unload(require)
+        box::unload(source)
+    })
+
+    expect_warning(box::use(mod/legacy/library), '.+library.+ inside a module')
+    expect_warning(box::use(mod/legacy/require), '.+require.+ inside a module')
+
+    expect_false(exists('source_test', envir = .GlobalEnv))
+    on.exit(rm(source_test, envir = .GlobalEnv))
+    expect_warning(box::use(mod/legacy/source), '.+source.+ inside a module')
+    expect_true(exists('source_test', envir = .GlobalEnv))
+})
+
+test_that('legacy function warning can be silenced', {
+    old_opts = options(box.warn.legacy = FALSE)
+    on.exit(options(old_opts))
+
+    expect_warning(box::use(mod/legacy/library), NA)
+    expect_warning(box::use(mod/legacy/require), NA)
+
+    expect_false(exists('source_test', envir = .GlobalEnv))
+    on.exit(rm(source_test, envir = .GlobalEnv))
+    expect_warning(box::use(mod/legacy/source), NA)
+    expect_true(exists('source_test', envir = .GlobalEnv))
+})